feat: add tenant_id check before deleting certificate

+++ c8y_test_core/assert_device_certificate.py
@@ -17,6 +17,8 @@ class AssertDeviceCertificate(AssertDevice):
         """Assert device certificate"""
         try:
             log.info("Removing device certificate. fingerprint=%s", fingerprint)
+            if not self.context.client.tenant_id:
+                raise ValueError(f"tenant_id is not set in api client")
             self.context.client.delete(
                 (
                     f"/tenant/tenants/{self.context.client.tenant_id}"
@@ -24,7 +26,7 @@ class AssertDeviceCertificate(AssertDevice):
                 ),
             )
         except KeyError as ex:
-            log.warning("Certificate does not exist, so nothing to delete. ex=%s", ex)
+            log.info("Certificate does not exist, so nothing to delete. ex=%s", ex)
         except ValueError as ex:
             log.error("Could not delete device certificate. ex=%s", ex)
             raise
